requirements.txt
absl-py==1.4.0
accelerate==0.20.3
addict==2.4.0
aiocache==0.12.0
aiohttp==3.9.1
aiosignal==1.3.1
anyio==3.7.0
APScheduler==3.10.4
arel==0.2.0
asttokens==2.2.1
async-asgi-testclient==1.4.11
async-timeout==4.0.3
attrs==22.1.0
audioread==3.0.1
Automat==22.10.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
av==12.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bidict==0.22.1
bitsandbytes==0.43.0
bitsrun==3.7.1
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.3
commonmark @ file:///Users/ktietz/demo/mc3/conda-bld/commonmark_1630649545323/work
constantly==23.10.4
contextlib2==21.6.0
contourpy==1.0.6
cryptography==42.0.5
cssselect==1.2.0
cvxpy==1.3.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///croot/cython_1676568029361/work
dataclasses==0.8
datasets==2.15.0
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
dictor==0.1.12
dill==0.3.7
diskcache==5.6.3
distinctipy==1.2.2
distlib==0.3.6
DNN-printer==0.0.2
docopt==0.6.2
docstring_parser==0.16
ecos==2.0.12
efficientnet-pytorch==0.7.1
einops==0.8.0
ephem==4.1.6
et-xmlfile==1.1.0
eval_type_backport==0.2.0
evaluate==0.4.2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.96.0
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
filelock==3.8.0
Flask==3.0.3
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
future @ file:///croot/future_1677599870788/work
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
giturlparse==0.10.0
google==2.0.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gpustat==1.1.1
grpcio==1.50.0
h11==0.14.0
httpcore==1.0.5
httptools==0.5.0
httpx==0.27.0
huggingface-hub==0.18.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.4
imageio==2.31.0
imageio-ffmpeg==0.5.1
importlib-metadata==6.7.0
importlib-resources==5.10.0
imutils==0.5.4
incremental==22.10.0
ipykernel==6.23.1
ipython==8.12.3
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
jiwer==3.0.4
jmespath==1.0.1
joblib==1.2.0
json-tricks==3.16.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.1
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab_pygments==0.3.0
kiwisolver==1.0.1
lazy_loader==0.2
librosa==0.10.2.post1
llvmlite==0.41.1
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
-e git+https://github.com/microsoft/GLIP.git@a5f302bfd4c5c67010e29f779e3b0bde94e89985#egg=maskrcnn_benchmark
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mmcls==0.25.0
# Editable install with no version control (mmcv-full==1.5.3)
-e /home/omnisky/anaconda3/envs/project-evodnn-system/lib/python3.8/site-packages
# Editable install with no version control (mmdet==2.25.0)
-e /home/omnisky/anaconda3/envs/project-evodnn-system/lib/python3.8/site-packages
mmsegmentation==0.25.0
moviepy==1.0.3
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.1
nlpaug==1.1.11
nltk==3.8.1
numba==0.58.1
numerize==0.12
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-ml-py==12.535.133
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.1.2
osqp==0.6.2.post8
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
pooch==1.8.2
portalocker==2.7.0
POT==0.9.3
prettytable==3.11.0
proglog==0.1.10
prompt-toolkit==3.0.38
Protego==0.3.0
protobuf==4.21.9
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
PTable==0.9.2
ptflops==0.7.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-trees==0.8.1
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ file:///home/conda/feedstock_root/build_artifacts/pycocotools_1667626621479/work
pycodestyle @ file:///croot/pycodestyle_1674267221883/work
pycparser==2.21
pydantic==1.10.8
pydicom==2.3.1
PyDispatcher==2.0.7
pydot==3.0.3
pygad==3.3.1
Pygments==2.13.0
pynrrd==0.4.3
pynvml==11.4.0
pyOpenSSL==24.1.0
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.21.1
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.5
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qdldl==0.1.5.post3
quadprog==0.1.11
queuelib==1.6.2
-e git+https://www.ghproxy.cn/https://github.com/ildoonet/pytorch-randaugment@48b8f509c4bbda93bbe733d98b3fd052b6e4c8ae#egg=RandAugment
rapidfuzz==3.9.6
ray==2.1.0
regex==2023.6.3
requests==2.28.1
requests-file==2.0.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.1
schema==0.7.5
scikit-image==0.21.0
scikit-learn==1.2.1
scikit-opt==0.6.6
scikit-video==1.1.11
scipy==1.9.3
Scrapy==2.11.1
scs==3.2.2
seaborn==0.13.0
sentencepiece==0.1.99
service-identity==24.1.0
Shapely==1.8.5.post1
shtab==1.7.1
SimpleITK==2.2.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
speedtest==0.0.1
stack-data==0.6.2
starlette==0.27.0
stringcase==1.2.0
supervisely==6.72.28
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.1
tinycss2==1.3.0
tldextract==5.1.1
tokenizers==0.13.3
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
torch==1.10.0
torch-pruning==1.2.5
torch-scatter==2.1.2
torchstat==0.0.7
torchvision==0.10.0
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets==5.9.0
transformers==4.33.0
trimesh==3.22.0
Twisted==24.3.0
typing_extensions==4.12.2
tyro==0.8.12
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
varname==0.11.1
virtualenv==20.16.7
w3lib==2.1.2
Wand==0.6.11
watchfiles==0.19.0
watchgod==0.6
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.6
wget==3.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.1
yarg==0.1.9
yarl==1.9.4
zipp==3.10.0
zope.interface==6.2
